+++ source/packages/@aws-accelerator/cdk-utils/lib/tagging-stacks.ts
@@ -62,6 +62,10 @@ export function addAcceleratorTags(node: IConstruct, partition: string, tags: IT
       if (resource.cfnResourceType === 'Custom::SsmPutParameterValue') {
         addTagsToPutSsmParameterResource(resource, tagsWithPrefix);
       }
+
+      if (resource.cfnResourceType === 'AWS::IAM::ManagedPolicy') {
+        setTagsForChildResources(resource, tagsWithPrefix);
+      }
     }
 
     if (resource instanceof cdk.CustomResourceProvider) {